modify type checking for server side and add defaultvalue for pet edit form

+++ src/app/app/actions/checkout-pet.ts
@@ -1,13 +1,20 @@
 "use server";
 
 import prisma from "@/lib/db";
+import { petIdSchema } from "@/lib/validation/pet-form-validation";
 import { revalidatePath } from "next/cache";
 
-export default async function checkoutPet(petId: string) {
+export default async function checkoutPet(petId: unknown) {
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return { message: "Invalid pet id" };
+  }
+
   try {
     await prisma.pet.delete({
       where: {
-        id: petId,
+        id: validatedPetId.data,
       },
     });
   } catch (error) {
